[edit] Rework the SignIn component to get color value
Deleted the axios importation and function as it is not used.
Replaced the object variable to store all values from the form by a simple variable to get only the color value.
Moved the handleChange function for the color to the SliderPicker component.
Added the route in the action attribute to specify the method.
Deleted the props passed to the Username, Email, Password and Submit components.
Added a hidden input field which gets the value of the SliderPicker component.